rename: UnzipFileList component
